refactor(codegen): Extract XHR and URL configurations
