+++ internal/code-gen/scripting/configuration/xhr_configuration.go
@@ -1,0 +1,28 @@
+package configuration
+
+import "github.com/gost-dom/code-gen/packagenames"
+
+func configureXHRSpecs(xhrModule *WebAPIConfig) {
+	xhrEventTarget := xhrModule.Type("XMLHttpRequestEventTarget")
+	xhrEventTarget.OverrideWrappedType = &GoType{
+		Package: packagenames.Events,
+		Name:    "EventTarget",
+	}
+
+	xhr := xhrModule.Type("XMLHttpRequest")
+
+	// TODO: Just need to support non-node objects
+	// xhr.SkipWrapper = true
+
+	xhr.MarkMembersAsNotImplemented(
+		"readyState",
+		"responseType",
+		"responseXML",
+	)
+	xhr.Method("open").SetCustomImplementation()
+	xhr.Method("upload").SetCustomImplementation()
+	xhr.Method("onreadystatechange").Ignore()
+
+	formData := xhrModule.Type("FormData")
+	formData.RunCustomCode = true
+}
